requirements.txt
Django==4.0.2
Markdown==3.3.6
django-filter==21.1
djangorestframework==3.13.1
djangorestframework_jsonapi==5.0.0
django-cors-headers==3.11.0
autopep8==1.6.0
python-dotenv==0.19.2
PyMySQL==1.0.2
